Correctly traverse pages based on user answers
The object responsible for traversing user answers can take a page
which tells it when to stop. If none is provided it would previously
take the last page in the `pages` array of the metadata. This was
perfectly fine before branching as the `pages` array was the thing
used to dictate the order of a form which always ended in a
Confirmation page.

However now branching uses a `flow` object system to dictate what a
user should see next based on their answers, the `pages` array order
no longer matters. The last page is simply the last page created by
the user.

Therefore we should traverse the user answers all the way to end of a
journey based upon their answers. In the instances of a page being
supplied we traverse up to and not including that page. If none is
provided then we go all the way until there is no longer the ability to
find another page. These instances are:

- end of the flow (i.e a confirmation page)
- an exit page
- a page with no default next which usually only happens if the form is
currently not been finished its creation by the Editor user