Add MoJ Forms global analytics
This injects the global analytics ID for the GA4 used by the MoJ Forms
team.

Since MoJ Forms will be using GA4 we do not want to overwrite a form
owners GA4 or vice versa. We only need to call the gtag function with
the additional measurement ID, as opposed to load the entire JS script
again.

Also changed `analytics_tag` to `measurement_id` as that is what Google
refers to it as.