Return both text and value when calling id on an AutocompleteItem
When the items for a component are passed into an autocomplete page
template `id` is used to set the 'value' on each select item. When a
user chooses and item and submits the answer it this value that gets
sent to the datastore.

However when we come to present the answer back to the user, either on
the autocomplete page when they go back to it or on a checkanswers page,
we want to have the text (aka name) as this is usually the human
readable form.

When we come to submit the form payload at that point we want to have
the 'value' since this is the thing that a Form Owner would actually
make use of.

Also add a countries fixture for use with testing.