Override to_json method on the component model
Calling to_json on a component objects injects all the component
attributes into the data-fb-content-data element for that component.
This is updated whenever a user makes changes to an interactable element
in the Editor, such as a component question or a section heading etc.

For autocomplete components we do not need to inject all the select
items into the data-fb-content-data element as they are not interactable
in the Editor. Therefore we can just remove them entirely before loading
the page.