Pass autocomplete_items through to the validator
We need the autocomplete items in the validators so we can compare
the user's answers to the items on the component.

We experimented with adding a module in the Runner and Editor apps
that would allow us access to the items, however, this left us with
the inability to sufficiently test the implementation. So we decided
to go with passing the items in from the answers controller.