Show the correct reference number dependant on the app
When in the Editor app, there is no need to generate a reference
number, so we want to use a placeholder.

When in the Runner, we want to generate a new reference number
every time.

`reference_number_enabled?` and `show_reference_number` are
methods that will be defined in the Editor and the Runner.