Fix Editor preview footer links
Changed the `editor_preview?` method to check the URL path includes
'preview'.

Previously, the method assumed all paths in the Editor in preview
would have 'preview' at the end.

This is not correct as the path appends the page slug to the path:
eg: 'service/:service_id/preview/:page_slug