Remove unwanted chars from  upload filename
We want to silently remove certain characters from file uploads.
This commit removes these certain characters.

We need to sanitise the input before removing the other special
characters. This is to first prevent any injection attacks.

We then need to remove the remaining '&gt;' or '&lt;' characters,
which have resulted from the sanitising.

Here, we have catered for a '\' char to be deleted and have a test
for this. However, with e2e testing the behaviour is chars before
the '\' is removed. This appeasr to be underlying behaviour in
the ActionDispatch class, which manipulates the 'filename' and
creates an accessor called 'original_filename'.
We receive the already manipulated original_filename in params.