Rename filename when extension is 'jfif' or 'jpg'
We have a user who relies on consolidating the uploaded file name
if the email attachments with the filenames in the users answers.

When a user uploads a file that has the extension of either 'jfif'
or 'jpg', the mimetype used is jpeg. This is because all these
extensions are all jpeg in the end.

This commit silently changes the filenames to ensure we use the
same extension that will be created when a jfif or jpg image is
uploaded.