Remove the body key from the Content page
By default, when creating a new content page, the first component
on the page looks like a content component but is the 'body'
on the schema.

Since this is not a content component, conditionals can not be
added to it.

Adjust the schema to force users to create a content component
when on the page.

When rendering the content page, we need to ensure that if there
the body is not the default '[Optional content]' we will need to
show it for backwards compatibility.