Rename EvaluateConditionals to EvaluateBranchConditionals
The branching conditionals return a page uuid for the next page
in the flow.

We can not re-use the same mechanism for conditional content
components as we will need to return the content component uuid
to be shown.

Rename these classes and methods to show they are specific to
branching.