Logic for conditional content
This introduces a new class `EvaluateContentConditionals` this
class takes a content component with conditionals and the user's
current answers.

It goes through each conditional in the component and compares
this to the user's answers, if there is a match, it will return
the component uuid (ie: the uuid of the content component to be shown).

Within this class we use the existing `Operators` class, which evaluates
the 'is', 'is not', 'contains', 'does not contain' operators.

We also use the existing `Conditional` class which evaluates the
expressions for the conditional.