Page Ids cannot be constants
Page Ids cannot be constants as they are generated from user-input and can be set to any arbitrary string (conforming to a few rules).

Before, this was a constant, and because the json-schema was an old version it was not validating this attribute. Now that it validates the attribute, it will fail if for instance a user creates an exit page and does not call it "exit", or deletes and creates again a confirmation page and does not call it "confirmation", etc.

By introducing in the page definition an "_id" of type string, that is inherited and validated by all types of pages (exit, confirmation, start, checkanswers, etc.) we solve the problem and in fact make the schemas more robust than before where some pages were not validating the presence/format of the "_id" attribute.