Update json-schema gem (#356)
* Update json-schema gem

Use latest `json-schema` version (4.1.1).

Since version 3.0.0, the gem also performs validation of `const` attributes.

Attribute `_type` (const) in _radios_, _checkboxes_ and _date_ components was being wrongly validated against a
  duplicated `_type` inherited from `definition.fieldset`. Removed this attribute from the fieldset definition.

`fb-editor` requires a few small changes in some specs in order to be able to use this new metadata-presenter gem. A separate PR will be raised in that repo.

Fixed warning "character class has '-' without escape" in a pattern regexp.

* Page Ids cannot be constants

Page Ids cannot be constants as they are generated from user-input and can be set to any arbitrary string (conforming to a few rules).

Before, this was a constant, and because the json-schema was an old version it was not validating this attribute. Now that it validates the attribute, it will fail if for instance a user creates an exit page and does not call it "exit", or deletes and creates again a confirmation page and does not call it "confirmation", etc.

By introducing in the page definition an "_id" of type string, that is inherited and validated by all types of pages (exit, confirmation, start, checkanswers, etc.) we solve the problem and in fact make the schemas more robust than before where some pages were not validating the presence/format of the "_id" attribute.

* No need for declaring the _id attribute

This attribute is in fact inherited via composition from `definition.data`

* Add missing page.exit to allowed page types