Attempt to mitigate unfiltered params edge case
As the `answers_controller` is the originator of the unfiltered params, which then are used in the `MetadataPresenter::PageAnswers`, I've introduced some handling and logging to attempt to permit ad-hoc the unfiltered params in the rare occasions these happen.

If we see this is enough we can remove/simplify code recently introduced in the runner.