Mitigate unfiltered params edge case and remove max files in number component (#360)
* Attempt to mitigate unfiltered params edge case

As the `answers_controller` is the originator of the unfiltered params, which then are used in the `MetadataPresenter::PageAnswers`, I've introduced some handling and logging to attempt to permit ad-hoc the unfiltered params in the rare occasions these happen.

If we see this is enough we can remove/simplify code recently introduced in the runner.

* Confirmed this safeguard is not needed

Seems like we never get to this point with unfiltered params. The safeguard in the `#upload_answer` method is the important one.

* Remove extraneous max_files validation in number

This validation was creeping in the number component but it is only relevant for multi file upload component.

* Bump gem version to 3.3.4

And reduce dependabot noise.