Update markup to wrap questions content components with stimulus functionality
Editable content areas are proving problematic.  This is due to the whole contents of the <editable-content>
custome element being inserted as editable content on initialization - meaning the up/down buttons were being included
within the editable component, and would be removed from the DOM on focus/unfocus.

To avoid this, we wrap the questions/content areas with a new div with all of the stimulus controller functionality.

Ediatble content areas still do not work correctly - as the custom element connectedCallback() fires each time they are moved
resulting in multiple content areas as they are moved around the page