Support localised error messages
https://trello.com/c/YohbrwI8

The error messages for each of the component validations are stored in json files.

This PR expands on a previous idea to have different `value` keys for each of the supported languages.

I've added a `value:cy` to all existing errors, that will need to be properly translated (there was only one existing translated error, in `error.required.json`).

Based on the current locale (coming from the metadata JSON and set on each request) the corresponding error message will be used on the runner.