Minimise localisation breaking changes
Some of these changes although minor require changes to the runner or the editor.

As this i18n work is still WIP and other devs are working in other features branching out from this repo `main`, it is best to keep any breaking changes out of it for now.

This way even if we were to cut a new gem version of the metadata presenter and use it in the other services, everything should keep working as usual.