Expose as_json method in address fieldset
This will be used in the runner as we will want to submit the address answer as JSON and not as a concatenated string as we were doing until now.
Updated tests.

This change is backwards compatible as current runners are calling `#to_a` which we still keep as it is used in the check your answers page.