Update upload component error message
https://trello.com/c/r3wfBU4f

The error message when a file is not selected in an uploader component needs to be custom and not to use the default used for the rest of components.

Instead of `Enter an answer for "Upload your best dog photo"` the error message will be now a generic `Choose a file to upload`.

Pending welsh translation.